Cleanup serialization code
