+++ Robust.Shared/Serialization/Manager/Attributes/CopyByRefAttribute.cs
@@ -2,6 +2,13 @@ using System;
 
 namespace Robust.Shared.Serialization.Manager.Attributes
 {
-    [AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct | AttributeTargets.Enum | AttributeTargets.Interface, Inherited = false)]
-    public class CopyByRefAttribute : Attribute {}
+    [AttributeUsage(
+        AttributeTargets.Class |
+        AttributeTargets.Struct |
+        AttributeTargets.Enum |
+        AttributeTargets.Interface,
+        Inherited = false)]
+    public class CopyByRefAttribute : Attribute
+    {
+    }
 }
