+++ Robust.Shared/Serialization/Manager/Attributes/ImplicitDataDefinitionForInheritorsAttribute.cs
@@ -3,5 +3,7 @@ using System;
 namespace Robust.Shared.Serialization.Manager.Attributes
 {
     [AttributeUsage(AttributeTargets.Class | AttributeTargets.Interface)]
-    public class ImplicitDataDefinitionForInheritorsAttribute : Attribute { }
+    public class ImplicitDataDefinitionForInheritorsAttribute : Attribute
+    {
+    }
 }
