+++ Robust.Shared/Serialization/Manager/Attributes/TypeSerializerAttribute.cs
@@ -5,5 +5,7 @@ namespace Robust.Shared.Serialization.Manager.Attributes
 {
     [AttributeUsage(AttributeTargets.Class, Inherited = false)]
     [MeansImplicitUse]
-    public class TypeSerializerAttribute : Attribute{}
+    public class TypeSerializerAttribute : Attribute
+    {
+    }
 }
