+++ Robust.Shared/Serialization/Manager/DataDefinition/FieldDefinition.cs
@@ -1,0 +1,45 @@
+﻿using System;
+using Robust.Shared.Serialization.Manager.Attributes;
+using Robust.Shared.Utility;
+
+namespace Robust.Shared.Serialization.Manager.DataDefinition
+{
+    public class FieldDefinition
+    {
+        public FieldDefinition(
+            DataFieldAttribute attr,
+            object? defaultValue,
+            AbstractFieldInfo fieldInfo,
+            AbstractFieldInfo backingField,
+            InheritanceBehaviour inheritanceBehaviour)
+        {
+            BackingField = backingField;
+            Attribute = attr;
+            DefaultValue = defaultValue;
+            FieldInfo = fieldInfo;
+            InheritanceBehaviour = inheritanceBehaviour;
+        }
+
+        public DataFieldAttribute Attribute { get; }
+
+        public object? DefaultValue { get; }
+
+        public InheritanceBehaviour InheritanceBehaviour { get; }
+
+        public AbstractFieldInfo BackingField { get; }
+
+        public AbstractFieldInfo FieldInfo { get; }
+
+        public Type FieldType => FieldInfo.FieldType;
+
+        public object? GetValue(object? obj)
+        {
+            return BackingField.GetValue(obj);
+        }
+
+        public void SetValue(object? obj, object? value)
+        {
+            BackingField.SetValue(obj, value);
+        }
+    }
+}
