+++ Robust.Shared/Serialization/Manager/DataDefinition/InheritanceBehaviour.cs
@@ -1,0 +1,9 @@
+﻿namespace Robust.Shared.Serialization.Manager.DataDefinition
+{
+    public enum InheritanceBehaviour : byte
+    {
+        Default,
+        Always,
+        Never
+    }
+}
