+++ Robust.Shared/Serialization/Manager/RequiredDataFieldNotProvidedException.cs
@@ -2,5 +2,7 @@ using System;
 
 namespace Robust.Shared.Serialization.Manager
 {
-    public class RequiredDataFieldNotProvidedException : Exception { }
+    public class RequiredDataFieldNotProvidedException : Exception
+    {
+    }
 }
