+++ Robust.Shared/Serialization/Manager/Result/DeserializedCollection.cs
@@ -20,7 +20,7 @@ namespace Robust.Shared.Serialization.Manager.Result
 
         public IEnumerable<DeserializationResult> Mappings { get; }
 
-        public override object? RawValue => Value;
+        public override object RawValue => Value;
 
         private Create CreateDelegate { get; }
