+++ Robust.Shared/Serialization/Manager/Result/DeserializedFieldEntry.cs
@@ -1,10 +1,10 @@
-﻿using System;
+﻿using Robust.Shared.Serialization.Manager.DataDefinition;
 
 namespace Robust.Shared.Serialization.Manager.Result
 {
     public class DeserializedFieldEntry
     {
-        public DeserializedFieldEntry(bool mapped, SerializationDataDefinition.InheritanceBehaviour inheritanceBehaviour, DeserializationResult? result = null)
+        public DeserializedFieldEntry(bool mapped, InheritanceBehaviour inheritanceBehaviour, DeserializationResult? result = null)
         {
             Mapped = mapped;
             Result = result;
@@ -12,20 +12,21 @@ namespace Robust.Shared.Serialization.Manager.Result
         }
 
         public bool Mapped { get; }
-        public SerializationDataDefinition.InheritanceBehaviour InheritanceBehaviour { get; }
+
+        public InheritanceBehaviour InheritanceBehaviour { get; }
 
         public DeserializationResult? Result { get; }
 
         public DeserializedFieldEntry PushInheritanceFrom(DeserializedFieldEntry fieldEntry)
         {
-            if(Mapped)
+            if (Mapped)
             {
-                if (InheritanceBehaviour == SerializationDataDefinition.InheritanceBehaviour.Always)
+                if (InheritanceBehaviour == InheritanceBehaviour.Always)
                 {
                     if (Result != null)
                     {
                         return fieldEntry.Result != null
-                            ? new(Mapped, InheritanceBehaviour, Result.PushInheritanceFrom(fieldEntry.Result))
+                            ? new DeserializedFieldEntry(Mapped, InheritanceBehaviour, Result.PushInheritanceFrom(fieldEntry.Result))
                             : Copy();
                     }
                     else
@@ -37,7 +38,7 @@ namespace Robust.Shared.Serialization.Manager.Result
                 return Copy();
             }
 
-            return InheritanceBehaviour == SerializationDataDefinition.InheritanceBehaviour.Never ? Copy() : fieldEntry.Copy();
+            return InheritanceBehaviour == InheritanceBehaviour.Never ? Copy() : fieldEntry.Copy();
         }
 
         public DeserializedFieldEntry Copy()
