+++ Robust.Shared/Serialization/Manager/SerializationManager.CustomSerializers.cs
@@ -1,4 +1,6 @@
-﻿using Robust.Shared.Serialization.Manager.Result;
+﻿using System;
+using System.Collections.Generic;
+using Robust.Shared.Serialization.Manager.Result;
 using Robust.Shared.Serialization.Markdown;
 using Robust.Shared.Serialization.Markdown.Validation;
 using Robust.Shared.Serialization.TypeSerializers.Interfaces;
@@ -7,22 +9,32 @@ namespace Robust.Shared.Serialization.Manager
 {
     public partial class SerializationManager
     {
-        private DeserializationResult ReadWithCustomTypeSerializer<T, TNode, TSerializer>(TNode node, ISerializationContext? context = null, bool skipHook = false)
-            where TSerializer : ITypeReader<T, TNode> where T : notnull where TNode : DataNode
+        private readonly Dictionary<Type, object> _customTypeSerializers = new();
+
+        private DeserializationResult ReadWithSerializer<T, TNode, TSerializer>(
+            TNode node,
+            ISerializationContext? context = null,
+            bool skipHook = false)
+            where TSerializer : ITypeReader<T, TNode>
+            where T : notnull
+            where TNode : DataNode
         {
             var serializer = (ITypeReader<T, TNode>)GetTypeSerializer(typeof(TSerializer));
             return serializer.Read(this, node, DependencyCollection, skipHook, context);
         }
 
-        private DataNode WriteWithCustomTypeSerializer<T, TSerializer>(T value,
-            ISerializationContext? context = null, bool alwaysWrite = false)
-            where TSerializer : ITypeWriter<T> where T : notnull
+        private DataNode WriteWithSerializer<T, TSerializer>(
+            T value,
+            ISerializationContext? context = null,
+            bool alwaysWrite = false)
+            where TSerializer : ITypeWriter<T>
+            where T : notnull
         {
             var serializer = (ITypeWriter<T>)GetTypeSerializer(typeof(TSerializer));
             return serializer.Write(this, value, alwaysWrite, context);
         }
 
-        private TCommon CopyWithCustomTypeSerializer<TCommon, TSource, TTarget, TSerializer>(
+        private TCommon CopyWithSerializer<TCommon, TSource, TTarget, TSerializer>(
             TSource source,
             TTarget target,
             bool skipHook,
@@ -36,7 +48,7 @@ namespace Robust.Shared.Serialization.Manager
             return serializer.Copy(this, source, target, skipHook, context);
         }
 
-        private ValidationNode ValidateWithCustomTypeSerializer<T, TNode, TSerializer>(
+        private ValidationNode ValidateWithSerializer<T, TNode, TSerializer>(
             TNode node,
             ISerializationContext? context)
             where T : notnull
