+++ Robust.Shared/Serialization/Manager/SerializationManager.cs
@@ -27,7 +27,7 @@ namespace Robust.Shared.Serialization.Manager
         private bool _initializing;
         private bool _initialized;
 
-        private readonly Dictionary<Type, SerializationDataDefinition> _dataDefinitions = new();
+        private readonly Dictionary<Type, DataDefinition.DataDefinition> _dataDefinitions = new();
         private readonly HashSet<Type> _copyByRefRegistrations = new();
 
         public IDependencyCollection DependencyCollection { get; private set; } = default!;
@@ -51,7 +51,6 @@ namespace Robust.Shared.Serialization.Manager
             InitializeFlagsAndConstants();
             InitializeTypeSerializers();
 
-            //var registrations = _reflectionManager.FindTypesWithAttribute<MeansDataDefinition>().ToHashSet();
             var registrations = new HashSet<Type>();
 
             foreach (var baseType in _reflectionManager.FindTypesWithAttribute<ImplicitDataDefinitionForInheritorsAttribute>())
@@ -87,7 +86,7 @@ namespace Robust.Shared.Serialization.Manager
                     continue;
                 }
 
-                _dataDefinitions.Add(type, new SerializationDataDefinition(type));
+                _dataDefinitions.Add(type, new DataDefinition.DataDefinition(type));
             }
 
             var error = new StringBuilder();
@@ -225,7 +224,7 @@ namespace Robust.Shared.Serialization.Manager
             ISerializationContext? context = null)
         {
             var method =
-                typeof(SerializationManager).GetRuntimeMethods().First(m => m.Name == nameof(ValidateWithCustomTypeSerializer))!.MakeGenericMethod(
+                typeof(SerializationManager).GetRuntimeMethods().First(m => m.Name == nameof(ValidateWithSerializer))!.MakeGenericMethod(
                     type, node.GetType(), typeSerializer);
             return (ValidationNode)method.Invoke(this, new object?[] {node, context})!;
         }
@@ -261,7 +260,7 @@ namespace Robust.Shared.Serialization.Manager
                 populateDefaultValues.PopulateDefaultValues();
             }
 
-            var res = dataDefinition.InvokePopulateDelegate(obj, definition.Mapping);
+            var res = dataDefinition.Populate(obj, definition.Mapping);
 
             if (!skipHook && res.RawValue is ISerializationHooks serializationHooksAfter)
             {
@@ -271,14 +270,14 @@ namespace Robust.Shared.Serialization.Manager
             return res;
         }
 
-        internal SerializationDataDefinition? GetDataDefinition(Type type)
+        internal DataDefinition.DataDefinition? GetDataDefinition(Type type)
         {
             if (_dataDefinitions.TryGetValue(type, out var dataDefinition)) return dataDefinition;
 
             return null;
         }
 
-        internal bool TryGetDataDefinition(Type type, [NotNullWhen(true)] out SerializationDataDefinition? dataDefinition)
+        internal bool TryGetDataDefinition(Type type, [NotNullWhen(true)] out DataDefinition.DataDefinition? dataDefinition)
         {
             dataDefinition = GetDataDefinition(type);
             return dataDefinition != null;
@@ -288,7 +287,6 @@ namespace Robust.Shared.Serialization.Manager
         {
             var underlyingType = type.EnsureNotNullableType();
 
-            // array
             if (underlyingType.IsArray)
             {
                 if (node is not SequenceDataNode sequenceDataNode) throw new InvalidNodeTypeException();
@@ -337,8 +335,6 @@ namespace Robust.Shared.Serialization.Manager
                 return new DeserializedValue(selfSerObj);
             }
 
-            //if (node is not MappingDataNode mappingDataNode) throw new InvalidNodeTypeException();
-
             if (underlyingType.IsInterface || underlyingType.IsAbstract)
             {
                  throw new InvalidOperationException($"Unable to create an instance of an interface or abstract type. Type: {underlyingType}");
@@ -358,12 +354,14 @@ namespace Robust.Shared.Serialization.Manager
 
             if (node is not MappingDataNode mappingDataNode)
             {
-                if(node is not ValueDataNode emptyValueDataNode || emptyValueDataNode.Value != "")
+                if (node is not ValueDataNode emptyValueDataNode || emptyValueDataNode.Value != string.Empty)
                     throw new ArgumentException($"No mapping node provided for type {type}");
-                mappingDataNode = new MappingDataNode(); //if we get an emptyValueDataNode we just use an empty mapping
+
+                // If we get an empty ValueDataNode we just use an empty mapping
+                mappingDataNode = new MappingDataNode();
             }
 
-            var res = dataDef.InvokePopulateDelegate(obj, mappingDataNode, this, context, skipHook);
+            var res = dataDef.Populate(obj, mappingDataNode, this, context, skipHook);
 
             if (!skipHook && res.RawValue is ISerializationHooks serHooks)
             {
@@ -398,7 +396,7 @@ namespace Robust.Shared.Serialization.Manager
         public DeserializationResult ReadWithTypeSerializer(Type type, Type typeSerializer, DataNode node, ISerializationContext? context = null,
             bool skipHook = false)
         {
-            var method = typeof(SerializationManager).GetRuntimeMethods().First(m => m.Name == nameof(ReadWithCustomTypeSerializer))!
+            var method = typeof(SerializationManager).GetRuntimeMethods().First(m => m.Name == nameof(ReadWithSerializer))!
                 .MakeGenericMethod(type, node.GetType(), typeSerializer);
             return (DeserializationResult) method.Invoke(this, new object?[] {node, context, skipHook})!;
         }
@@ -424,7 +422,6 @@ namespace Robust.Shared.Serialization.Manager
                 return new ValueDataNode(convertible.ToString(CultureInfo.InvariantCulture));
             }
 
-            // array
             if (underlyingType.IsArray)
             {
                 var sequenceNode = new SequenceDataNode();
@@ -471,7 +468,7 @@ namespace Robust.Shared.Serialization.Manager
                 throw new ArgumentException($"Supplied value does not fit with data definition of {type}.");
             }
 
-            var newMapping = dataDef.InvokeSerializeDelegate(value, this, context, alwaysWrite);
+            var newMapping = dataDef.Serialize(value, this, context, alwaysWrite);
             mapping = mapping.Merge(newMapping);
 
             return mapping;
@@ -480,9 +477,10 @@ namespace Robust.Shared.Serialization.Manager
         public DataNode WriteWithTypeSerializer(Type type, Type typeSerializer, object? value, bool alwaysWrite = false,
             ISerializationContext? context = null)
         {
-            if (type.IsNullable() && value == null) return new MappingDataNode(); //todo just return null
+            // TODO Serialization: just return null
+            if (type.IsNullable() && value == null) return new MappingDataNode();
 
-            var method = typeof(SerializationManager).GetRuntimeMethods().First(m => m.Name == nameof(WriteWithCustomTypeSerializer))!
+            var method = typeof(SerializationManager).GetRuntimeMethods().First(m => m.Name == nameof(WriteWithSerializer))!
                 .MakeGenericMethod(type, typeSerializer);
             return (DataNode) method.Invoke(this, new object?[] {value, context, alwaysWrite})!;
         }
@@ -508,7 +506,6 @@ namespace Robust.Shared.Serialization.Manager
                     $"Source and target do not match. Source ({sourceType}) is value type? {sourceType.IsValueType}. Target ({targetType}) is value type? {targetType.IsValueType}");
             }
 
-            // array
             if (sourceType.IsArray && targetType.IsArray)
             {
                 var sourceArray = (Array) source;
@@ -564,7 +561,7 @@ namespace Robust.Shared.Serialization.Manager
                 throw new InvalidOperationException($"No data definition found for type {commonType} when copying");
             }
 
-            target = dataDef.InvokeCopyDelegate(source, target, this, context);
+            target = dataDef.Copy(source, target, this, context);
 
             if (!skipHook && target is ISerializationHooks afterHooks)
             {
@@ -598,7 +595,7 @@ namespace Robust.Shared.Serialization.Manager
                 throw new InvalidOperationException($"Could not find common type in {nameof(CopyWithTypeSerializer)}!");
             }
 
-            var method = typeof(SerializationManager).GetRuntimeMethods().First(m => m.Name == nameof(CopyWithCustomTypeSerializer))!
+            var method = typeof(SerializationManager).GetRuntimeMethods().First(m => m.Name == nameof(CopyWithSerializer))!
                 .MakeGenericMethod(commonType, source.GetType(), target.GetType(), typeSerializer);
             return method.Invoke(this, new object?[] {source, target, skipHook, context});
         }
