+++ Robust.Shared/Serialization/TypeSerializers/Implementations/EntitySerializer.cs
@@ -25,8 +25,6 @@ namespace Robust.Shared.Serialization.TypeSerializers.Implementations
             }
 
             var entity = dependencies.Resolve<IEntityManager>().GetEntity(uid);
-
-            // TODO Paul what type to return here
             return new DeserializedValue<IEntity>(entity);
         }
 
@@ -34,7 +32,6 @@ namespace Robust.Shared.Serialization.TypeSerializers.Implementations
             IDependencyCollection dependencies,
             ISerializationContext? context = null)
         {
-            // TODO Paul should we be checking entity exists here
             return EntityUid.TryParse(node.Value, out var uid) &&
                    uid.IsValid() &&
                    dependencies.Resolve<IEntityManager>().EntityExists(uid)
