+++ Robust.Shared/Utility/SpecificPropertyInfo.cs
@@ -1,0 +1,117 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
+using System.Linq;
+using System.Reflection;
+
+namespace Robust.Shared.Utility
+{
+    public class SpecificPropertyInfo : AbstractFieldInfo
+    {
+        public override string Name { get; }
+
+        public readonly PropertyInfo PropertyInfo;
+        internal override MemberInfo MemberInfo => PropertyInfo;
+
+        public override Type FieldType => PropertyInfo.PropertyType;
+        public override Type? DeclaringType => PropertyInfo.DeclaringType;
+
+        public SpecificPropertyInfo(PropertyInfo propertyInfo)
+        {
+            Name = propertyInfo.Name;
+            PropertyInfo = propertyInfo;
+        }
+
+        public override object? GetValue(object? obj) => PropertyInfo.GetValue(obj);
+        public override void SetValue(object? obj, object? value) => PropertyInfo.SetValue(obj, value);
+
+        public override T? GetCustomAttribute<T>() where T : class
+        {
+            if (PropertyInfo.TryGetCustomAttribute(out T? attribute))
+            {
+                return attribute;
+            }
+
+            return GetBackingField()?.GetCustomAttribute<T>();
+        }
+
+        public override IEnumerable<T> GetCustomAttributes<T>()
+        {
+            return PropertyInfo.GetCustomAttributes<T>()
+                .Concat(GetBackingField()?.GetCustomAttributes<T>() ?? Enumerable.Empty<T>());
+        }
+
+        public override bool HasCustomAttribute<T>()
+        {
+            return PropertyInfo.HasCustomAttribute<T>() ||
+                   (GetBackingField()?.HasCustomAttribute<T>() ?? false);
+        }
+
+        public override bool TryGetCustomAttribute<T>([NotNullWhen(true)] out T? attribute) where T : class
+        {
+            return (attribute = GetCustomAttribute<T>()) != null;
+        }
+
+        public override bool IsBackingField()
+        {
+            return false;
+        }
+
+        public override bool HasBackingField()
+        {
+            return DeclaringType?.HasBackingField(Name) ?? false;
+        }
+
+        public override SpecificFieldInfo? GetBackingField()
+        {
+            return DeclaringType?.GetBackingField(Name);
+        }
+
+        public override bool TryGetBackingField([NotNullWhen(true)] out SpecificFieldInfo? field)
+        {
+            return (field = GetBackingField()) != null;
+        }
+
+        public bool IsMostOverridden(Type type)
+        {
+            if (DeclaringType == type)
+            {
+                return true;
+            }
+
+            var setBase = PropertyInfo.SetMethod?.GetBaseDefinition();
+            var getBase = PropertyInfo.GetMethod?.GetBaseDefinition();
+            var currentType = type;
+            var relevantProperties = type.GetAllProperties().Where(p => p.Name == PropertyInfo.Name).ToList();
+            while (currentType != null)
+            {
+                foreach (var property in relevantProperties)
+                {
+                    if(property.DeclaringType != currentType) continue;
+
+                    if (setBase != null && setBase == property.SetMethod?.GetBaseDefinition())
+                    {
+                        return property == PropertyInfo;
+                    }
+
+                    if (getBase != null && getBase == property.GetMethod?.GetBaseDefinition())
+                    {
+                        return property == PropertyInfo;
+                    }
+                }
+
+                currentType = currentType.BaseType;
+            }
+
+            return false;
+        }
+
+        public static implicit operator PropertyInfo(SpecificPropertyInfo f) => f.PropertyInfo;
+        public static explicit operator SpecificPropertyInfo(PropertyInfo f) => new(f);
+
+        public override string? ToString()
+        {
+            return PropertyInfo.ToString();
+        }
+    }
+}
