+++ Robust.Shared/Utility/TypeHelpers.cs
@@ -119,143 +119,6 @@ namespace Robust.Shared.Utility
             }
         }
 
-        public static bool TryGenericReadOnlyCollectionType(Type type, [NotNullWhen(true)] out Type? listType)
-        {
-            if (!type.GetTypeInfo().IsGenericType)
-            {
-                listType = default;
-                return false;
-            }
-
-            var baseGeneric = type.GetGenericTypeDefinition();
-            var isList = baseGeneric == typeof(IReadOnlyCollection<>) || baseGeneric == typeof(IReadOnlyList<>);
-
-            if (isList)
-            {
-                listType = type.GetGenericArguments()[0];
-                return true;
-            }
-
-            listType = default;
-            return false;
-        }
-
-        public static bool TryGenericListType(Type type, [NotNullWhen(true)] out Type? listType)
-        {
-            var isList = type.GetTypeInfo().IsGenericType && type.GetGenericTypeDefinition() == typeof(List<>);
-
-            if (isList)
-            {
-                listType = type.GetGenericArguments()[0];
-                return true;
-            }
-
-            listType = default;
-            return false;
-        }
-
-        public static bool TryGenericReadOnlyDictType(Type type, [NotNullWhen(true)] out Type? keyType, [NotNullWhen(true)] out Type? valType)
-        {
-            var isDict = type.GetTypeInfo().IsGenericType && type.GetGenericTypeDefinition() == typeof(IReadOnlyDictionary<,>);
-
-            if (isDict)
-            {
-                var genArgs = type.GetGenericArguments();
-                keyType = genArgs[0];
-                valType = genArgs[1];
-                return true;
-            }
-
-            keyType = default;
-            valType = default;
-            return false;
-        }
-
-        public static bool TryGenericReadDictType(Type type, [NotNullWhen(true)] out Type? keyType,
-            [NotNullWhen(true)] out Type? valType, [NotNullWhen(true)] out Type? dictType)
-        {
-            if (TryGenericDictType(type, out keyType, out valType))
-            {
-                // Pass through the type directly if it's Dictionary<K,V>.
-                // Since that's more efficient.
-                dictType = type;
-                return true;
-            }
-
-            if (TryGenericReadOnlyDictType(type, out keyType, out valType))
-            {
-                // If it's IReadOnlyDictionary<K,V> we need to make a Dictionary<K,V> type to use to deserialize.
-                dictType = typeof(Dictionary<,>).MakeGenericType(keyType, valType);
-                return true;
-            }
-
-            dictType = default;
-            return false;
-        }
-
-        public static bool TryGenericDictType(Type type, [NotNullWhen(true)] out Type? keyType, [NotNullWhen(true)] out Type? valType)
-        {
-            var isDict = type.GetTypeInfo().IsGenericType && type.GetGenericTypeDefinition() == typeof(Dictionary<,>);
-
-            if (isDict)
-            {
-                var genArgs = type.GetGenericArguments();
-                keyType = genArgs[0];
-                valType = genArgs[1];
-                return true;
-            }
-
-            keyType = default;
-            valType = default;
-            return false;
-        }
-
-        public static bool TryGenericSortedDictType(Type type, [NotNullWhen(true)] out Type? keyType,
-            [NotNullWhen(true)] out Type? valType)
-        {
-            var isDict = type.GetTypeInfo().IsGenericType && type.GetGenericTypeDefinition() == typeof(SortedDictionary<,>);
-
-            if (isDict)
-            {
-                var genArgs = type.GetGenericArguments();
-                keyType = genArgs[0];
-                valType = genArgs[1];
-                return true;
-            }
-
-            keyType = default;
-            valType = default;
-            return false;
-        }
-
-        public static bool TryGenericHashSetType(Type type, [NotNullWhen(true)] out Type? setType)
-        {
-            var isSet = type.GetTypeInfo().IsGenericType && type.GetGenericTypeDefinition() == typeof(HashSet<>);
-
-            if (isSet)
-            {
-                setType = type.GetGenericArguments()[0];
-                return true;
-            }
-
-            setType = default;
-            return false;
-        }
-
-        public static bool TryGenericSortedSetType(Type type, [NotNullWhen(true)] out Type? setType)
-        {
-            var isSet = type.GetTypeInfo().IsGenericType && type.GetGenericTypeDefinition() == typeof(SortedSet<>);
-
-            if (isSet)
-            {
-                setType = type.GetGenericArguments()[0];
-                return true;
-            }
-
-            setType = default;
-            return false;
-        }
-
         public static IEnumerable<AbstractFieldInfo> GetAllPropertiesAndFields(this Type type)
         {
             foreach (var field in type.GetAllFields())
@@ -326,204 +189,6 @@ namespace Robust.Shared.Utility
             where T : Attribute
         {
             return (attribute = memberInfo.GetCustomAttribute<T>()) != null;
-        }
-    }
-
-    public abstract class AbstractFieldInfo
-    {
-        public abstract string Name { get; }
-        internal abstract MemberInfo MemberInfo { get; }
-        public abstract Type FieldType { get; }
-        public abstract Type? DeclaringType { get; }
-
-        public abstract object? GetValue(object? obj);
-        public abstract void SetValue(object? obj, object? value);
-
-        public abstract T? GetCustomAttribute<T>() where T : Attribute;
-        public abstract IEnumerable<T> GetCustomAttributes<T>() where T : Attribute;
-        public abstract bool HasCustomAttribute<T>() where T : Attribute;
-        public abstract bool TryGetCustomAttribute<T>([NotNullWhen(true)] out T? attribute) where T : Attribute;
-        public abstract bool IsBackingField();
-        public abstract bool HasBackingField();
-        public abstract SpecificFieldInfo? GetBackingField();
-        public abstract bool TryGetBackingField([NotNullWhen(true)] out SpecificFieldInfo? field);
-    }
-
-    public class SpecificFieldInfo : AbstractFieldInfo
-    {
-        public override string Name { get; }
-
-        public readonly FieldInfo FieldInfo;
-        internal override MemberInfo MemberInfo => FieldInfo;
-
-        public override Type FieldType => FieldInfo.FieldType;
-        public override Type? DeclaringType => FieldInfo.DeclaringType;
-
-        public SpecificFieldInfo(FieldInfo fieldInfo)
-        {
-            Name = fieldInfo.Name;
-            FieldInfo = fieldInfo;
-        }
-
-        public override object? GetValue(object? obj) => FieldInfo.GetValue(obj);
-        public override void SetValue(object? obj, object? value) => FieldInfo.SetValue(obj, value);
-
-        public override T? GetCustomAttribute<T>() where T : class
-        {
-            return FieldInfo.GetCustomAttribute<T>();
-        }
-
-        public override IEnumerable<T> GetCustomAttributes<T>()
-        {
-            return FieldInfo.GetCustomAttributes<T>();
-        }
-
-        public override bool HasCustomAttribute<T>()
-        {
-            return FieldInfo.HasCustomAttribute<T>();
-        }
-
-        public override bool TryGetCustomAttribute<T>([NotNullWhen(true)] out T? attribute) where T : class
-        {
-            return FieldInfo.TryGetCustomAttribute(out attribute);
-        }
-
-        public override bool IsBackingField()
-        {
-            return FieldInfo.IsBackingField();
-        }
-
-        public override bool HasBackingField()
-        {
-            return false;
-        }
-
-        public override SpecificFieldInfo? GetBackingField()
-        {
-            return null;
-        }
-
-        public override bool TryGetBackingField([NotNullWhen(true)] out SpecificFieldInfo? field)
-        {
-            field = null;
-            return false;
-        }
-
-        public static implicit operator FieldInfo(SpecificFieldInfo f) => f.FieldInfo;
-        public static explicit operator SpecificFieldInfo(FieldInfo f) => new(f);
-
-        public override string? ToString()
-        {
-            return FieldInfo.ToString();
-        }
-    }
-
-    public class SpecificPropertyInfo : AbstractFieldInfo
-    {
-        public override string Name { get; }
-
-        public readonly PropertyInfo PropertyInfo;
-        internal override MemberInfo MemberInfo => PropertyInfo;
-
-        public override Type FieldType => PropertyInfo.PropertyType;
-        public override Type? DeclaringType => PropertyInfo.DeclaringType;
-
-        public SpecificPropertyInfo(PropertyInfo propertyInfo)
-        {
-            Name = propertyInfo.Name;
-            PropertyInfo = propertyInfo;
-        }
-
-        public override object? GetValue(object? obj) => PropertyInfo.GetValue(obj);
-        public override void SetValue(object? obj, object? value) => PropertyInfo.SetValue(obj, value);
-
-        public override T? GetCustomAttribute<T>() where T : class
-        {
-            if (PropertyInfo.TryGetCustomAttribute(out T? attribute))
-            {
-                return attribute;
-            }
-
-            return GetBackingField()?.GetCustomAttribute<T>();
-        }
-
-        public override IEnumerable<T> GetCustomAttributes<T>()
-        {
-            return PropertyInfo.GetCustomAttributes<T>()
-                .Concat(GetBackingField()?.GetCustomAttributes<T>() ?? Enumerable.Empty<T>());
-        }
-
-        public override bool HasCustomAttribute<T>()
-        {
-            return PropertyInfo.HasCustomAttribute<T>() ||
-                   (GetBackingField()?.HasCustomAttribute<T>() ?? false);
-        }
-
-        public override bool TryGetCustomAttribute<T>([NotNullWhen(true)] out T? attribute) where T : class
-        {
-            return (attribute = GetCustomAttribute<T>()) != null;
-        }
-
-        public override bool IsBackingField()
-        {
-            return false;
-        }
-
-        public override bool HasBackingField()
-        {
-            return DeclaringType?.HasBackingField(Name) ?? false;
-        }
-
-        public override SpecificFieldInfo? GetBackingField()
-        {
-            return DeclaringType?.GetBackingField(Name);
-        }
-
-        public override bool TryGetBackingField([NotNullWhen(true)] out SpecificFieldInfo? field)
-        {
-            return (field = GetBackingField()) != null;
-        }
-
-        public bool IsMostOverridden(Type type)
-        {
-            if (DeclaringType == type)
-            {
-                return true;
-            }
-
-            var setBase = PropertyInfo.SetMethod?.GetBaseDefinition();
-            var getBase = PropertyInfo.GetMethod?.GetBaseDefinition();
-            var currentType = type;
-            var relevantProperties = type.GetAllProperties().Where(p => p.Name == PropertyInfo.Name).ToList();
-            while (currentType != null)
-            {
-                foreach (var property in relevantProperties)
-                {
-                    if(property.DeclaringType != currentType) continue;
-
-                    if (setBase != null && setBase == property.SetMethod?.GetBaseDefinition())
-                    {
-                        return property == PropertyInfo;
-                    }
-
-                    if (getBase != null && getBase == property.GetMethod?.GetBaseDefinition())
-                    {
-                        return property == PropertyInfo;
-                    }
-                }
-
-                currentType = currentType.BaseType;
-            }
-
-            return false;
-        }
-
-        public static implicit operator PropertyInfo(SpecificPropertyInfo f) => f.PropertyInfo;
-        public static explicit operator SpecificPropertyInfo(PropertyInfo f) => new(f);
-
-        public override string? ToString()
-        {
-            return PropertyInfo.ToString();
         }
     }
 }
